requirements.txt
phidata
python-dotenv
packaging
fastapi
uvicorn
typer
pgvector
psycopg[binary]
psycopg2-binary
pypdf
duckduckgo-search
sqlalchemy
openai
google-generativeai